+++ src/br/com/folha/model/login/BeanOperadorLogado.java
@@ -16,14 +16,14 @@ public class BeanOperadorLogado {
     private String nome;
     private String cpf;
     private String privilegio;
-    private int seqOperador;
-    private int seqPrivilegio;
+    private long seqOperador;
+    private long seqPrivilegio;
     
 
     public BeanOperadorLogado() {
     }
 
-    public BeanOperadorLogado(String nome, String cpf, String privilegio, int seqOperador, int seqPrivilegio) {
+    public BeanOperadorLogado(String nome, String cpf, String privilegio, long seqOperador, long seqPrivilegio) {
         this.nome = nome;
         this.cpf = cpf;
         this.privilegio = privilegio;
@@ -55,19 +55,19 @@ public class BeanOperadorLogado {
         this.privilegio = privilegio;
     }
 
-    public int getSeqOperador() {
+    public long getSeqOperador() {
         return seqOperador;
     }
 
-    public void setSeqOperador(int seqOperador) {
+    public void setSeqOperador(long seqOperador) {
         this.seqOperador = seqOperador;
     }
 
-    public int getSeqPrivilegio() {
+    public long getSeqPrivilegio() {
         return seqPrivilegio;
     }
 
-    public void setSeqPrivilegio(int seqPrivilegio) {
+    public void setSeqPrivilegio(long seqPrivilegio) {
         this.seqPrivilegio = seqPrivilegio;
     }
